Added user agent flag, and pop-up alerts bypass flag. Enable file downloads.
Specify the user agent using --useragent
Bypass pop-up alerts using -alert
Can now download files, and will be put into their respective folders.

+++ grim-scraper.py
@@ -13,7 +13,26 @@ import sys
 from validator_collection import validators, checkers
 import csv
 import argparse
+from selenium.webdriver.common.alert import Alert
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.common.exceptions import TimeoutException
+from selenium.webdriver.common.keys import Keys
 
+from selenium.webdriver.support.ui import WebDriverWait
+
+from selenium.webdriver.common.action_chains import ActionChains
+
+def check_alerts(driver):
+    try:
+
+        WebDriverWait(driver, 1).until(EC.alert_is_present(),'Timed out')
+        print("Alert has showed up...")
+        alerts = driver.switch_to.alert
+        alerts.accept()
+        print("Alert dismissed...")
+    except TimeoutException:
+        print("No alert found...")
+            
 
 
 def extract_root_domain(url):
@@ -25,17 +44,26 @@ def extract_root_domain(url):
     return top
 
 
-def initialize_driver(headless_mode):
+def initialize_driver(headless_mode, useragent, root_domain):
     options = webdriver.ChromeOptions()
+    path = os.path.dirname(os.path.abspath(__file__))
+    download_dir = path+"/"+root_domain+"/downloads"
+
+    prefs = {
+    "download.default_directory" : download_dir,
+    'safebrowsing.enabled': True
+    }      
+    options.add_experimental_option("prefs", prefs)
+
+    options.add_argument("user-agent="+useragent)
     options.headless = headless_mode
     driver = webdriver.Chrome(ChromeDriverManager().install(), options=options)
     print("Driver has been initialized")
-    return driver
+    return driver, download_dir
 
 
 def take_screenshot(driver, url, root_domain):
     screenshot_loc = root_domain+"/screenshot.png"
-    Path(root_domain).mkdir(parents=True, exist_ok=True)
     driver.save_screenshot(screenshot_loc)
     print("Saved screenshot of " +url+" in " + screenshot_loc)
 
@@ -85,7 +113,7 @@ def save_to_csv(url, status_code, content_type):
     data = [url, status_code]
 
 
-def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filetype):
+def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filetype, alert, download_dir):
     print("Saving the resources...")
     try:
         for url in urls:
@@ -99,7 +127,9 @@ def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filet
             if ( (filetype_str.find(filetype) == -1) and (filetype != "*")):
                 continue
             driver.get(url)
-            time.sleep(1)
+            time.sleep(2)
+            if alert:
+                check_alerts(driver)
             src = driver.page_source
             file_path = urlparse(url).path
             file_name = os.path.basename(file_path)
@@ -110,8 +140,6 @@ def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filet
                 file_name = "index"
 
             scrape_dir = url_root_domain +"/"+ domain_path
-
-
             
 
             if len(file_name) > 250:
@@ -138,6 +166,7 @@ def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filet
 
             full_path = scrape_dir + file_name
 
+
             if os.path.isfile(full_path):
                 new_path = scrape_dir + str(i)+"_"+file_name
                 i=i+1
@@ -145,12 +174,17 @@ def reap(driver, urls, url_root_domain, logs, all_resource, urls_filetype, filet
                     print(full_path + " exists, renaming to " +new_path)
                 full_path = new_path
 
-
-            f = open(full_path, "w")
-            f.write(src)
-            f.close()
-            if logs:
-                print("Saved " + url + " in " + full_path)
+            #Check if file is inside /downloads. If so move to the proper directory
+            downloaded = download_dir+"/"+file_name
+            if os.path.isfile(downloaded):
+                os.rename(downloaded, full_path)
+                print(file_name + " exists in " +downloaded+ ", moved to "+ full_path)
+            else: 
+                f = open(full_path, "w")
+                f.write(src)
+                f.close()
+                if logs:
+                    print("Saved " + url + " in " + full_path)
             
     except Exception as e:
         print(e)
@@ -165,9 +199,11 @@ def main():
         parser = argparse.ArgumentParser()
         parser.add_argument('--url', required=True, help="Specify the URL")
         parser.add_argument('--filetype', required=False, help="Specify the resource filetype to save")
+        parser.add_argument('--useragent', required=False, help="Specify the user agent")
         parser.add_argument('-headless', action='store_true', help="Run in headless mode")
         parser.add_argument('-log', action='store_true', help="Output logs")
         parser.add_argument('-all', action='store_true', help="Save all resources found in HTTP request/response")
+        parser.add_argument('-alert', action='store_true', help="Accept pop up alert")
         args = parser.parse_args()
         url = args.url
 
@@ -183,6 +219,13 @@ def main():
         else:
             filetype = "*"
             print("Will save resources of any filetype...")
+
+        if args.useragent:
+            useragent = args.useragent
+            print("Will use the uger agent "+useragent+"...")
+        else:
+            useragent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/113.0.0.0 Safari/537.36"
+            print("Will use the default user agent "+useragent+"...")
         
 
         if args.headless:
@@ -217,17 +260,30 @@ def main():
             else:
                 print("Will only save main resources with filetype "+filetype+"...")
 
-        driver = initialize_driver(headless_mode)
+        if args.alert:
+            alert = True
+            print("Will accept pop alert...")
+        else:
+            alert = False
+
+        Path(root_domain).mkdir(parents=True, exist_ok=True)
+        print("Made a directory " +root_domain)
+        driver, download_dir = initialize_driver(headless_mode, useragent, root_domain)
         driver.get(url)
-        time.sleep(5)
+        time.sleep(2)
+
+        if alert:
+            check_alerts(driver)
+            
 
         take_screenshot(driver, url, root_domain)
 
         urls, urls_filetype = http_responses(driver, url, root_domain, logs)
 
-        reap(driver, urls, root_domain, logs, all_resource, urls_filetype, filetype)
+        reap(driver, urls, root_domain, logs, all_resource, urls_filetype, filetype, alert, download_dir)
 
     except Exception as e:
+        print(e)
         print("Error... exiting")
         sys.exit()
